Caches - resctructure and use Generics
